+++ apps/station/lib/station/markers/markers.ex
@@ -1,7 +1,7 @@
 defmodule Tankste.Station.Markers do
 
   alias Tankste.Station.Markers.Marker
-  alias Tankste.Station.Stations
+  alias Tankste.Station.StationInfos
   alias Tankste.Station.OpenTimes
   alias Tankste.Station.Repo
 
@@ -11,36 +11,61 @@ defmodule Tankste.Station.Markers do
   def gen_by_boundary(boundary) do
     reduced_boundary = min_max_boundary(boundary)
 
-    scope_stations = Stations.list(status: "available", boundary: reduced_boundary |> boundary_with_padding())
-      |> Repo.preload([:open_times, station_areas: [area: [:holidays]]])
-      |> Enum.map(fn s -> %{s | is_open: OpenTimes.is_open(s)} end)
-      |> Repo.preload(:prices)
+    scope_station_infos = StationInfos.list(boundary: reduced_boundary |> boundary_with_padding())
+      |> Enum.sort_by(fn si -> si.priority end, :desc)
+      |> Enum.uniq_by(fn si -> si.station_id end)
+      |> Repo.preload([:station, :open_times, station_areas: [area: [:holidays]]])
+      |> Enum.filter(fn si -> si.station.status == "available" end)
+      |> Enum.map(fn si -> %{si | is_open: OpenTimes.is_open(si)} end)
+      |> Repo.preload(station: [:prices])
+      |> Enum.map(fn si ->
+        prices = Enum.sort_by(si.station.prices, fn p -> p.priority end, :desc)
+        |> Enum.uniq_by(fn p -> p.type end)
 
-      comparing_stations = scope_stations
-        |> Enum.filter(fn s -> s.is_open end)
+        Map.put(si, :station, Map.put(si.station, :prices, prices))
+      end)
 
-      scope_stations
-      |> Enum.filter(fn s -> in_boundary({s.location_latitude, s.location_longitude}, reduced_boundary) end)
-      |> Enum.map(fn s -> gen_marker(s, comparing_stations) end)
+      comparing_station_infos = scope_station_infos
+        |> Enum.filter(fn si -> si.is_open end)
+
+        scope_station_infos
+      |> Enum.filter(fn si -> in_boundary({si.location_latitude, si.location_longitude}, reduced_boundary) end)
+      |> Enum.map(fn si -> gen_marker(si, comparing_station_infos) end)
   end
 
   def gen_by_station_id(station_id) do
-    station = Stations.get(station_id, status: "available")
-    station = station
+    station_info = StationInfos.list(station_id: station_id)
+      |> Enum.sort_by(fn si -> si.priority end, :desc)
+      |> Repo.preload(:station)
+      |> Enum.filter(fn si -> si.station.status == "available" end)
+      |> Enum.at(0)
       |> Repo.preload([:open_times, station_areas: [area: [:holidays]]])
-      |> Map.put(:is_open, OpenTimes.is_open(station))
-      |> Repo.preload(:prices)
 
-    scope_boundary = [{station.location_latitude, station.location_longitude}, {station.location_latitude, station.location_longitude}]
+    station_info = station_info
+      |> Map.put(:is_open, OpenTimes.is_open(station_info))
+      |> Repo.preload(station: [:prices])
+
+    station_info = Map.put(station_info, :station, Map.put(station_info.station, :prices, Enum.sort_by(station_info.station.prices, fn p -> p.priority end) |> Enum.uniq_by(fn p -> p.type end)))
+
+    scope_boundary = [{station_info.location_latitude, station_info.location_longitude}, {station_info.location_latitude, station_info.location_longitude}]
       |> boundary_with_padding()
 
-    comparing_stations = Stations.list(status: "available", boundary: scope_boundary)
-      |> Repo.preload([:open_times, station_areas: [area: [:holidays]]])
-      |> Enum.map(fn s -> %{s | is_open: OpenTimes.is_open(s)} end)
+    comparing_station_infos = StationInfos.list(boundary: scope_boundary)
+      |> Enum.sort_by(fn si -> si.priority end, :desc)
+      |> Enum.uniq_by(fn si -> si.station_id end)
+      |> Repo.preload([:station, :open_times, station_areas: [area: [:holidays]]])
+      |> Enum.filter(fn si -> si.station.status == "available" end)
+      |> Enum.map(fn si -> %{si | is_open: OpenTimes.is_open(si)} end)
       |> Enum.filter(fn s -> s.is_open end)
-      |> Repo.preload(:prices)
+      |> Repo.preload(station: [:prices])
+      |> Enum.map(fn si ->
+        prices = Enum.sort_by(si.station.prices, fn p -> p.priority end, :desc)
+          |> Enum.uniq_by(fn p -> p.type end)
 
-      gen_marker(station, comparing_stations)
+        Map.put(si, :station, Map.put(si.station, :prices, prices))
+      end)
+
+      gen_marker(station_info, comparing_station_infos)
   end
 
   defp min_max_boundary(boundary) do
@@ -63,46 +88,46 @@ defmodule Tankste.Station.Markers do
   defp latitude_param({latitude, _}), do: latitude
   defp longitude_param({_, longitude}), do: longitude
 
-  defp gen_marker(station, comparing_stations) do
-    near_stations = comparing_stations
-      |> Enum.filter(fn s -> Geocalc.within?(@station_distance_comparing_meters, [s.location_longitude, s.location_latitude], [station.location_longitude, station.location_latitude]) end)
+  defp gen_marker(station_info, comparing_station_infos) do
+    near_station_infos = comparing_station_infos
+      |> Enum.filter(fn si -> Geocalc.within?(@station_distance_comparing_meters, [si.location_longitude, si.location_latitude], [station_info.location_longitude, station_info.location_latitude]) end)
 
-    case station.is_open do
+    case station_info.is_open do
       true ->
         %Marker{
-          id: station.id,
-          station_id: station.id,
-          label: station.brand || station.name,
-          latitude: station.location_latitude,
-          longitude: station.location_longitude,
-          e5_price: get_price(station, "e5"),
-          e5_price_comparison: get_price_comparison(station, "e5", near_stations),
-          e10_price: get_price(station, "e10"),
-          e10_price_comparison: get_price_comparison(station, "e10", near_stations),
-          diesel_price: get_price(station, "diesel"),
-          diesel_price_comparison: get_price_comparison(station, "diesel", near_stations),
-          currency: station.currency
+          id: station_info.station_id,
+          station_id: station_info.station_id,
+          label: station_info.brand || station_info.name,
+          latitude: station_info.location_latitude,
+          longitude: station_info.location_longitude,
+          e5_price: get_price(station_info, "e5"),
+          e5_price_comparison: get_price_comparison(station_info, "e5", near_station_infos),
+          e10_price: get_price(station_info, "e10"),
+          e10_price_comparison: get_price_comparison(station_info, "e10", near_station_infos),
+          diesel_price: get_price(station_info, "diesel"),
+          diesel_price_comparison: get_price_comparison(station_info, "diesel", near_station_infos),
+          currency: station_info.currency
         }
       _ ->
         %Marker{
-          id: station.id,
-          station_id: station.id,
-          label: station.brand || station.name,
-          latitude: station.location_latitude,
-          longitude: station.location_longitude,
+          id: station_info.station_id,
+          station_id: station_info.station_id,
+          label: station_info.brand || station_info.name,
+          latitude: station_info.location_latitude,
+          longitude: station_info.location_longitude,
           e5_price: nil,
           e5_price_comparison: "not_available",
           e10_price: nil,
           e10_price_comparison: "not_available",
           diesel_price: nil,
           diesel_price_comparison: "not_available",
-          currency: station.currency
+          currency: station_info.currency
         }
     end
   end
 
-  defp get_price(station, type) do
-    case Enum.find(station.prices, fn p -> p.type == type end) do
+  defp get_price(station_info, type) do
+    case Enum.find(station_info.station.prices, fn p -> p.type == type end) do
     nil ->
       nil
     price ->
@@ -110,13 +135,13 @@ defmodule Tankste.Station.Markers do
     end
   end
 
-  defp get_price_comparison(station, type, near_stations) do
-    near_prices = near_stations
-      |> Enum.flat_map(fn s -> s.prices end)
+  defp get_price_comparison(station_info, type, near_station_infos) do
+    near_prices = near_station_infos
+      |> Enum.flat_map(fn si -> si.station.prices end)
       |> Enum.filter(fn p -> p.type == type end)
       |> Enum.map(fn p -> p.price end)
 
-    case get_price(station, type) do
+    case get_price(station_info, type) do
       nil ->
         "not_available"
       price_value ->
@@ -124,8 +149,8 @@ defmodule Tankste.Station.Markers do
           |> Enum.min()
 
         cond do
-          min_price + get_price_medium_threshold_value(station.currency) >= price_value -> "cheap"
-          min_price + get_price_expensive_threshold_value(station.currency) >= price_value -> "medium"
+          min_price + get_price_medium_threshold_value(station_info.currency) >= price_value -> "cheap"
+          min_price + get_price_expensive_threshold_value(station_info.currency) >= price_value -> "medium"
           true -> "expensive"
         end
     end