+++ apps/station/lib/station/station_areas/station_area.ex
@@ -4,8 +4,8 @@ defmodule Tankste.Station.StationAreas.StationArea do
   import Ecto.Changeset
 
   # TODO: ecto supports n-2-n tables, refactore afterwards and use the benefits of this
-  schema "station_areas" do
-    belongs_to :station, Tankste.Station.Stations.Station
+  schema "station_info_areas" do
+    belongs_to :station_info, Tankste.Station.StationInfos.StationInfo
     belongs_to :area, Tankste.Station.Areas.Area
 
     timestamps()
@@ -13,7 +13,7 @@ defmodule Tankste.Station.StationAreas.StationArea do
 
   def changeset(open_time, attrs) do
     open_time
-    |> cast(attrs, [:station_id, :area_id])
-    |> validate_required([:station_id, :area_id])
+    |> cast(attrs, [:station_info_id, :area_id])
+    |> validate_required([:station_info_id, :area_id])
   end
 end