+++ apps/station/priv/repo/migrations/20240303223714_add_price_currency.exs
@@ -2,7 +2,6 @@ defmodule Tankste.Station.Repo.Migrations.AddPriceCurrency do
   use Ecto.Migration
 
   def up do
-
     alter table(:stations) do
       add :currency, :string, null: false, default: "eur"
     end