+++ apps/station_web/lib/station_web/controllers/station_controller.ex
@@ -2,12 +2,15 @@ defmodule Tankste.StationWeb.StationController do
   use Tankste.StationWeb, :controller
 
   alias Tankste.Station.Stations
+  alias Tankste.Station.StationInfos
   alias Tankste.Station.OpenTimes
 
   plug :load_station when action in [:show]
 
   def show(conn, _params) do
-    render(conn, "show.json", station: station(conn))
+    conn
+    |> put_view(Tankste.StationWeb.StationInfoView)
+    |> render("show.json", station_info: station_info(conn))
   end
 
   def load_station(conn, opts \\ []) do
@@ -23,13 +26,26 @@ defmodule Tankste.StationWeb.StationController do
         |> halt
 
       station ->
-        station = station
-          |> Map.put(:is_open, OpenTimes.is_open(station))
+        case StationInfos.get_by_station_id(station_id) do
+          nil ->
+            conn
+            |> put_status(:not_found)
+            |> put_view(ErrorView)
+            |> render("404.json")
+            |> halt
 
-        conn
-        |> assign(:station, station)
+          station_info ->
+            station_info = station_info
+            |> Map.put(:is_open, OpenTimes.is_open(station_info))
+
+            conn
+            |> assign(:station, station)
+            |> assign(:station_info, station_info)
+        end
     end
   end
 
   def station(conn), do: Map.get(conn.assigns, :station)
+
+  def station_info(conn), do: Map.get(conn.assigns, :station_info)
 end