+++ apps/station_web/lib/station_web/controllers/station_price_controller.ex
@@ -1,7 +1,7 @@
 defmodule Tankste.StationWeb.StationPriceController do
   use Tankste.StationWeb, :controller
 
-  import Tankste.StationWeb.StationController, only: [load_station: 2, station: 1]
+  import Tankste.StationWeb.StationController, only: [load_station: 2, station_info: 1]
 
   alias Tankste.Station.Prices
 
@@ -9,9 +9,11 @@ defmodule Tankste.StationWeb.StationPriceController do
 
   def index(conn, %{"station_id" => station_id}) do
     prices = Prices.list(station_id: station_id)
-    prices = case station(conn).is_open do
+    prices = case station_info(conn).is_open do
         true ->
           prices
+          |> Enum.sort_by(fn p -> p.priority end, :desc)
+          |> Enum.uniq_by(fn p -> p.type end)
         _ ->
           []
       end