+++ apps/station_web/lib/station_web/views/open_time_view.ex
@@ -4,7 +4,7 @@ defmodule Tankste.StationWeb.OpenTimeView do
   def render("open_time.json", %{open_time: open_time}) do
     %{
       "id" => open_time.id,
-      "stationId" => open_time.station_id,
+      "stationInfoId" => open_time.station_info_id,
       "originId" => open_time.origin_id,
       "day" => open_time.day,
       "startTime" => open_time.start_time,