+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 
 ## Upcoming version ##
 
+- 🛠 Allow multiple data origins for stations, prices & related data ([#7](https://github.com/tankste/backend/issues/7))
 - 🌟 Add currency ([#8](https://github.com/tankste/backend/issues/8))
 
 ## 1.0.0 (03.03.2024) ##